requirements.txt
Appium-Python-Client==2.8.1
attrs==22.2.0
more-itertools==9.0.0
packaging==23.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.2.2
pytest-html==3.2.0
pytest-metadata==2.0.4
PyYAML==6.0
selenium==4.8.0
six==1.16.0
urllib3==1.26.14
wcwidth==0.2.6
h11~=0.14.0
pip~=21.2.4
wheel~=0.37.0
Jinja2~=3.1.2
trio~=0.22.0
outcome~=1.2.0
sniffio~=1.3.0
exceptiongroup~=1.1.0
sortedcontainers~=2.4.0
idna~=3.4
setuptools~=58.0.4
iniconfig~=2.0.0
certifi~=2022.12.7
zipp~=3.13.0
future~=0.18.2
wsproto~=1.2.0
requests~=2.28.2